requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.3
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.1
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.28.2
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
webdriver-manager==4.0.1
pyopenssl==23.3.0
cryptography==41.0.5
pillow==10.1.0
tqdm==4.66.1 